+++ src/unified/factory/index.ts
@@ -1,2 +1,3 @@
 export * from './createEcdhCurve';
 export * from './createSignatureCurve';
+export * from './createSignatureCurveRngDisallowed';